+++ tests/integration/setup/test_multi_setup_preger.py
@@ -1,7 +1,6 @@
 import math
 
 from pyoma2.algorithms import SSI_MS, pLSCF_MS
-
 from src.pyoma2.setup import MultiSetup_PreGER